Using TimeLeft Dictionary and refactoring get_current_time
